# Conflicts:
#	x/cpp/crdt/json.gen.h
#	x/cpp/crdt/types.gen.h
#	x/go/crdt/codec.gen.go
#	x/go/crdt/pb/crdt.pb.go
#	x/go/crdt/pb/crdt.proto
#	x/py/x/crdt/types_gen.py

+++ x/cpp/crdt/json.gen.h
@@ -22,7 +22,7 @@ namespace x::crdt {
 inline ID ID::parse(x::json::Parser parser) {
     return ID{
         .replica = parser.field<std::uint32_t>("replica"),
-        .counter = parser.field<std::uint64_t>("counter"),
+        .counter = parser.field<std::uint32_t>("counter"),
     };
 }
